+++ app/src/main/java/com/sixtyninefourtwenty/imdefinitelysober/internal/CircularBuffer.kt
@@ -1,4 +1,4 @@
-package com.sixtyninefourtwenty.imdefinitelysober
+package com.sixtyninefourtwenty.imdefinitelysober.internal
 
 import java.io.Serializable
